+++ TagsCloud/Spiral.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloud
 {
 	internal class Spiral
 	{
