+++ TagsCloud/TagsCloud.csproj
@@ -8,7 +8,7 @@
     <ProjectGuid>{0A4BBA84-B783-4A44-9159-DCF9AB0E1629}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TagsCloudVisualization</RootNamespace>
+    <RootNamespace>TagsCloud</RootNamespace>
     <AssemblyName>TagsCloud</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
